requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
paho-mqtt==1.6.1
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.0
httpx==0.25.2
